pyproject.toml
[project]
name = "edusolve"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.9.1",
    "asyncio==3.4.3",
    "docx2pdf==0.1.8",
    "faiss-cpu==1.7.4",
    "google-generativeai==0.3.2",
    "groq==0.4.2",
    "httpx==0.27.2",
    "huggingface-hub==0.20.3",
    "langchain==0.1.0",
    "langchain-community==0.0.10",
    "numpy==1.26.2",
    "pandas==2.1.4",
    "pillow==10.1.0",
    "pymupdf==1.23.8",
    "python-docx==1.1.0",
    "python-dotenv==1.0.0",
    "reportlab==4.0.7",
    "sentence-transformers==2.2.2",
    "streamlit==1.52.0",
    "tenacity==8.2.3",
]
